refactor(ui): split SkillsSection into sources/install/list components
- settings/skillsApi.ts: skill list/update/marketplace types + fetches
- settings/SkillsSourcesEditor.tsx: skills.sources array editor (SYNC_ALL_KEY)
- settings/SkillsInstallSearch.tsx: marketplace search dropdown (owns its
  query state and the INSTALL_MENU_LIMIT cap)
- settings/SkillsInstalledList.tsx: installed list with switch/update/delete
- shared icons moved to settings/settingsIcons.tsx; apiSend reused from
  settings/settingsApi.ts
Markup and testids unchanged; auto-discovery fieldset stays in the section.
